laid out new data processing file structure
